+++ include/bitwuzla/c/bitwuzla.h
@@ -2092,12 +2092,14 @@ BitwuzlaTerm bitwuzla_mk_fp_from_real(BitwuzlaTermManager* tm,
  * @note Given rounding mode may be an arbitrary, non-value rounding mode term.
  *       If it is a value, the returned term will be a floating-point value,
  *       else a non-value floating-point term.
+ * @note Given denominator must be non-zero.
  *
  * @param tm The term manager instance.
  * @param sort The sort of the value.
  * @param rm The rounding mode.
  * @param num The decimal string representing the numerator.
- * @param den The decimal string representing the denominator.
+ * @param den The decimal string representing the denominator, must not
+ *            represent zero.
  *
  * @return A floating-point representation of the given rational string. If
  *         `rm` is of kind BITWUZLA_KIND_VALUE the floating-point will be of